fix: move static args json utils
